+++ Project_TOY_client_c_sharp_unity/Assets/Scripts/Managers/Contents/DataManager.cs
@@ -4,12 +4,12 @@ using System.Collections.Generic;
 public class DataManager
 {
     // 외부에서 Managers.Data.StatDict[1] 로 바로 접근 가능
-    public Dictionary<int, Stat> StatDict { get; private set; } = new Dictionary<int, Stat>();
+    public Dictionary<ulong, Stat> StatDict { get; private set; } = new Dictionary<ulong, Stat>();
 
     public void Init()
     {
         // 1. JSON 로드 (Resources/Data/StatData.json)
-        StatDict = LoadJson<StatData, int, Stat>("StatData").MakeDict();
+        StatDict = LoadJson<StatData, ulong, Stat>("StatData").MakeDict();
     }
 
     // 제네릭을 이용해 어떤 타입의 데이터든 로드하는 범용 함수
